Move custom Adapters to adapters.py
